requirements.txt
pandas
requests
rapidfuzz
numpy
sqlalchemy
psycopg2-binary
scikit-learn
xgboost
rapidfuzz
requests